06-agent-engine/greeting_agent/requirements.txt
google-cloud-aiplatform[agent_engines, adk]